// src/generated/products.ts
// @generated by `npm run protocol:generate`; do not edit.
export type Product = { id: string, revision: number, name: string, description: string, currency: string | null, price: number | null, stock: number | null, skus: Array<ProductSku>, images: Array<string>, createdAt: string, updatedAt: string, };

export type ProductSku = { code: string, name: string, price: number | null, stock: number | null, };

export type ProductInlineImage = { name: string, base64: string, };

export type ProductSaveRequest = { id: string | null, revision: number | null, name: string, description: string, currency: string | null, price: number | null, stock: number | null, skus: Array<ProductSku>, images: Array<string>, inlineImages: Array<ProductInlineImage>, };